+++ src/main/java/logger/ConsoleLogger.java
@@ -13,14 +13,7 @@ public class ConsoleLogger implements LoggerClass{
     private final Logger consoleLogger = LoggerFactory.getLogger(ConsoleLogger.class);
 
     @Override
-    public void waitForInput() {
-        try (Scanner scanner = new Scanner(System.in)) {
-            System.out.println("Waiting for new lines. Key in Ctrl+D to exit.");
-            while (true) {
-                String line = scanner.nextLine();
-                consoleLogger.info(line);
-            }
-        } catch (IllegalStateException | NoSuchElementException e) {
-        }
+    public void log(String message){
+        consoleLogger.info(message);
     }
 }
