+++ src/main/java/logger/FileConsoleLogger.java
@@ -11,14 +11,9 @@ import java.util.logging.Level;
 @Singleton
 public class FileConsoleLogger implements LoggerClass{
     private final Logger consoleFileLogger = LoggerFactory.getLogger(FileConsoleLogger.class);
+
     @Override
-    public void waitForInput() {
-        try (Scanner scanner = new Scanner(System.in)) {
-            System.out.println("Waiting for new lines. Key in Ctrl+D to exit.");
-            while (true) {
-                consoleFileLogger.info(scanner.nextLine());
-            }
-        } catch (IllegalStateException | NoSuchElementException e) {
-        }
+    public void log(String message){
+        consoleFileLogger.info(message);
     }
 }
