+++ src/main/java/logger/FileLogger.java
@@ -11,14 +11,9 @@ import java.util.logging.Level;
 @Singleton
 public class FileLogger implements LoggerClass{
     private final Logger fileLogger = LoggerFactory.getLogger(FileLogger.class);
+
     @Override
-    public void waitForInput() {
-        try (Scanner scanner = new Scanner(System.in)) {
-            System.out.println("Waiting for new lines. Key in Ctrl+D to exit.");
-            while (true) {
-                fileLogger.info(scanner.nextLine());
-            }
-        } catch (IllegalStateException | NoSuchElementException e) {
-        }
+    public void log(String message){
+        fileLogger.info(message);
     }
 }
