+++ src/main/java/module/Module.java
@@ -1,13 +1,10 @@
 package module;
 
 import com.google.inject.AbstractModule;
-import com.google.inject.Key;
 import logger.ConsoleLogger;
 import logger.FileConsoleLogger;
 import logger.FileLogger;
 import logger.LoggerClass;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 public class Module extends AbstractModule {
     private String param;
@@ -30,7 +27,5 @@ public class Module extends AbstractModule {
                 bind(LoggerClass.class).to(FileConsoleLogger.class);
                 break;
         }
-
-
     }
 }
