+++ src/main/java/util/Application.java
@@ -1,0 +1,30 @@
+package util;
+
+import com.google.inject.Guice;
+import com.google.inject.Inject;
+import com.google.inject.Injector;
+import logger.LoggerClass;
+import module.Module;
+
+import java.util.NoSuchElementException;
+import java.util.Scanner;
+
+public class Application {
+    @Inject
+    private LoggerClass logger;
+
+    public static void main(String[] args) {
+        final Injector injector = Guice.createInjector(new Module(args[0]));
+        injector.getInstance(Application.class).waitForInput();
+    }
+
+    public void waitForInput() {
+        try (Scanner scanner = new Scanner(System.in)) {
+            System.out.println("Waiting for new lines. Key in Ctrl+D to exit.");
+            while (true) {
+                logger.log(scanner.nextLine());
+            }
+        } catch (IllegalStateException | NoSuchElementException e) {
+        }
+    }
+}
